MODORDERS-251 # Refactoring
